Refactor recoverer for more clarity
