+++ lib/core/src/recover/handlers/mod.rs
@@ -1,0 +1,9 @@
+mod handle_chain_receive_swap;
+mod handle_chain_send_swap;
+mod handle_receive_swap;
+mod handle_send_swap;
+
+pub(crate) use self::handle_chain_receive_swap::ChainReceiveSwapHandler;
+pub(crate) use self::handle_chain_send_swap::ChainSendSwapHandler;
+pub(crate) use self::handle_receive_swap::ReceiveSwapHandler;
+pub(crate) use self::handle_send_swap::SendSwapHandler;
