+++ lib/core/src/recover/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod handlers;
 pub(crate) mod model;
 pub(crate) mod recoverer;
