+++ lib/core/src/sdk.rs
@@ -291,6 +291,7 @@ impl LiquidSdkBuilder {
             persister.clone(),
             swapper.clone(),
             liquid_chain_service.clone(),
+            recoverer.clone(),
         );
 
         let receive_swap_handler = ReceiveSwapHandler::new(
