+++ lib/core/src/send_swap.rs
@@ -19,8 +19,8 @@ use crate::persist::model::PaymentTxDetails;
 use crate::prelude::{PaymentTxData, PaymentType, Swap};
 use crate::recover::recoverer::Recoverer;
 use crate::swapper::Swapper;
+use crate::utils;
 use crate::wallet::OnchainWallet;
-use crate::{ensure_sdk, utils};
 use crate::{
     error::PaymentError,
     model::{PaymentState, Transaction as SdkTransaction},
@@ -35,6 +35,7 @@ pub(crate) struct SendSwapHandler {
     swapper: Arc<dyn Swapper>,
     chain_service: Arc<dyn LiquidChainService>,
     subscription_notifier: broadcast::Sender<String>,
+    recoverer: Arc<Recoverer>,
 }
 
 #[sdk_macros::async_trait]
@@ -55,6 +56,7 @@ impl SendSwapHandler {
         persister: Arc<Persister>,
         swapper: Arc<dyn Swapper>,
         chain_service: Arc<dyn LiquidChainService>,
+        recoverer: Arc<Recoverer>,
     ) -> Self {
         let (subscription_notifier, _) = broadcast::channel::<String>(30);
         Self {
@@ -64,6 +66,7 @@ impl SendSwapHandler {
             swapper,
             chain_service,
             subscription_notifier,
+            recoverer,
         }
     }
 
@@ -113,12 +116,19 @@ impl SendSwapHandler {
                     }
                     None => {
                         debug!("The claim tx was a script path spend (non-cooperative claim)");
-                        let preimage = self
-                            .get_preimage_from_script_path_claim_spend(&swap)
-                            .await?;
-                        self.validate_send_swap_preimage(id, &swap.invoice, &preimage)
-                            .await?;
-                        self.update_swap_info(id, Complete, Some(&preimage), None, None)?;
+                        let mut swaps = vec![Swap::Send(swap.clone())];
+                        self.recoverer.recover_from_onchain(&mut swaps).await?;
+
+                        let Swap::Send(s) = swaps[0].clone() else {
+                            return Err(anyhow!("Expected a Send swap"));
+                        };
+                        self.update_swap(s)?;
+                        // let preimage = self
+                        //     .get_preimage_from_script_path_claim_spend(&swap)
+                        //     .await?;
+                        // self.validate_send_swap_preimage(id, &swap.invoice, &preimage)
+                        //     .await?;
+                        // self.update_swap_info(id, Complete, Some(&preimage), None, None)?;
                     }
                 }
 
@@ -369,66 +379,6 @@ impl SendSwapHandler {
         self.swapper
             .claim_send_swap_cooperative(send_swap, claim_tx_details, &output_address)
             .await?;
-        Ok(())
-    }
-
-    pub(crate) async fn get_preimage_from_script_path_claim_spend(
-        &self,
-        swap: &SendSwap,
-    ) -> Result<String, PaymentError> {
-        info!("Retrieving preimage from non-cooperative claim tx");
-
-        let id = &swap.id;
-        let swap_script = swap.get_swap_script()?;
-        let swap_script_pk = swap_script
-            .to_address(self.config.network.into())?
-            .script_pubkey();
-        debug!("Found Send Swap swap_script_pk: {swap_script_pk:?}");
-
-        // Get tx history of the swap script (lockup address)
-        let history: Vec<_> = self
-            .chain_service
-            .get_script_history(&swap_script_pk)
-            .await?;
-
-        // We expect at most 2 txs: lockup and maybe the claim
-        ensure_sdk!(
-            history.len() <= 2,
-            PaymentError::Generic {
-                err: format!("Lockup address history for Send Swap {id} has more than 2 txs")
-            }
-        );
-
-        match history.get(1) {
-            None => Err(PaymentError::Generic {
-                err: format!("Send Swap {id} has no claim tx"),
-            }),
-            Some(claim_tx_entry) => {
-                let claim_tx_id = claim_tx_entry.txid;
-                let claim_tx = self
-                    .chain_service
-                    .get_transactions(&[claim_tx_id])
-                    .await
-                    .map_err(|e| anyhow!("Failed to fetch claim txs {claim_tx_id:?}: {e}"))?
-                    .first()
-                    .cloned()
-                    .ok_or(anyhow!("Claim tx not found for Send swap {id}"))?;
-
-                Ok(Recoverer::get_send_swap_preimage_from_claim_tx(
-                    id, &claim_tx,
-                )?)
-            }
-        }
-    }
-
-    async fn validate_send_swap_preimage(
-        &self,
-        swap_id: &str,
-        invoice: &str,
-        preimage: &str,
-    ) -> Result<(), PaymentError> {
-        utils::verify_payment_hash(preimage, invoice)?;
-        info!("Preimage is valid for Send Swap {swap_id}");
         Ok(())
     }
