+++ lib/core/src/swapper/mod.rs
@@ -61,6 +61,7 @@ pub trait Swapper: Send + Sync {
     async fn get_submarine_pairs(&self) -> Result<Option<SubmarinePair>, PaymentError>;
 
     /// Get a submarine swap's preimage
+    #[allow(dead_code)]
     async fn get_submarine_preimage(&self, swap_id: &str) -> Result<String, PaymentError>;
 
     /// Get send swap claim tx details which includes the preimage as a proof of payment.
