+++ lib/core/src/test_utils/send_swap.rs
@@ -5,12 +5,13 @@ use std::sync::Arc;
 use crate::{
     model::{Config, Signer},
     persist::Persister,
+    recover::recoverer::Recoverer,
     send_swap::SendSwapHandler,
 };
 use anyhow::Result;
 
 use super::{
-    chain::MockLiquidChainService,
+    chain::{MockBitcoinChainService, MockLiquidChainService},
     swapper::MockSwapper,
     wallet::{MockSigner, MockWallet},
 };
@@ -18,9 +19,18 @@ use super::{
 pub(crate) fn new_send_swap_handler(persister: Arc<Persister>) -> Result<SendSwapHandler> {
     let config = Config::testnet(None);
     let signer: Arc<Box<dyn Signer>> = Arc::new(Box::new(MockSigner::new()?));
-    let onchain_wallet = Arc::new(MockWallet::new(signer)?);
+    let onchain_wallet = Arc::new(MockWallet::new(signer.clone())?);
     let swapper = Arc::new(MockSwapper::default());
     let chain_service = Arc::new(MockLiquidChainService::new());
+    let liquid_chain_service = Arc::new(MockLiquidChainService::new());
+    let bitcoin_chain_service = Arc::new(MockBitcoinChainService::new());
+    let recoverer = Arc::new(Recoverer::new(
+        signer.slip77_master_blinding_key()?,
+        swapper.clone(),
+        onchain_wallet.clone(),
+        liquid_chain_service.clone(),
+        bitcoin_chain_service.clone(),
+    )?);
 
     Ok(SendSwapHandler::new(
         config,
@@ -28,5 +38,6 @@ pub(crate) fn new_send_swap_handler(persister: Arc<Persister>) -> Result<SendSwa
         persister,
         swapper,
         chain_service,
+        recoverer,
     ))
 }
